requirements.txt
asgiref==3.3.1
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
psycopg2-binary==2.8.6
python-magic==0.4.25
pytz==2020.4
sqlparse==0.4.1
